requirement.txt
absl-py==0.11.0
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1605075753540/work
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.3
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1600454382015/work
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
fastapi==0.61.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h11==0.11.0
h5py==2.10.0
idna==2.10
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1602263269022/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1605455388979/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==2.11.2
json5 @ file:///home/conda/feedstock_root/build_artifacts/json5_1600692310011/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1602551949684/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1598486169312/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1605735024330/work
jupyterlab==2.2.9
jupyterlab-server @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_server_1593951277307/work
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1602267317906/work
matplotlib==3.3.3
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1605115839832/work
natsort==7.0.1
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1602732862338/work
notebook @ file:///Users/runner/miniforge3/conda-bld/notebook_1594379841067/work
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandocfilters==1.4.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.0.1
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1605543085815/work
prompt-toolkit==1.0.15
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1603558917696/work
pyparsing==2.4.7
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1605115594146/work
python-dateutil==2.8.1
pyzmq==19.0.2
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
Send2Trash==1.5.0
simplegeneric==0.8.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
soupsieve==2.0.1
starlette==0.13.6
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado @ file:///Users/runner/miniforge3/conda-bld/terminado_1605116921917/work
testpath==0.4.4
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1604105030757/work
tqdm==4.53.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work